Move fixtures to conftest.py
This is the advice given by
https://docs.pytest.org/en/latest/fixture.html#conftest-py-sharing-fixture-functions.
